refactor: update CLI implementation and adjust ignore files
